move params to separate mod under simdrive
