+++ fastsim-core/src/simdrive/params.rs
@@ -1,0 +1,186 @@
+use super::*;
+
+#[serde_api]
+#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
+#[non_exhaustive]
+#[serde(deny_unknown_fields)]
+#[cfg_attr(feature = "pyo3", pyclass(module = "fastsim", subclass, eq))]
+/// Solver parameters
+pub struct SimParams {
+    #[serde(default = "SimParams::def_ach_speed_max_iter")]
+    /// max number of iterations allowed in setting achieved speed when trace
+    /// cannot be achieved
+    pub ach_speed_max_iter: u32,
+    #[serde(default = "SimParams::def_ach_speed_tol")]
+    /// tolerance in change in speed guess in setting achieved speed when trace
+    /// cannot be achieved
+    pub ach_speed_tol: si::Ratio,
+    #[serde(default = "SimParams::def_ach_speed_solver_gain")]
+    /// Newton method gain for setting achieved speed
+    pub ach_speed_solver_gain: f64,
+    // TODO: plumb this up to actually do something
+    /// When implemented, this will set the tolerance on how much trace miss
+    /// is allowed
+    #[serde(default = "SimParams::def_trace_miss_tol")]
+    pub trace_miss_tol: TraceMissTolerance,
+    #[serde(default = "SimParams::def_trace_miss_opts")]
+    pub trace_miss_opts: TraceMissOptions,
+    #[serde(default = "SimParams::def_trace_miss_correct_max_steps")]
+    /// the maximum number of steps in which to re-rendezvous with reference
+    /// trace after a trace miss. Note: this field only applies when
+    /// trace_miss_opts is set to TraceMissOptions::Correct. Note: must
+    /// be 2 or greater. Defaults to 6.
+    pub trace_miss_correct_max_steps: u32,
+    /// whether to use FASTSim-2 style air density
+    #[serde(default = "SimParams::def_f2_const_air_density")]
+    pub f2_const_air_density: bool,
+    /// if true, vehicle is totally inactive except for thermal models
+    pub ambient_thermal_soak: bool,
+}
+
+#[pyo3_api]
+impl SimParams {
+    #[staticmethod]
+    #[pyo3(name = "default")]
+    fn default_py() -> Self {
+        Self::default()
+    }
+}
+
+impl SimParams {
+    fn def_ach_speed_max_iter() -> u32 {
+        Self::default().ach_speed_max_iter
+    }
+    fn def_ach_speed_tol() -> si::Ratio {
+        Self::default().ach_speed_tol
+    }
+    fn def_ach_speed_solver_gain() -> f64 {
+        Self::default().ach_speed_solver_gain
+    }
+    fn def_trace_miss_tol() -> TraceMissTolerance {
+        Self::default().trace_miss_tol
+    }
+    fn def_trace_miss_opts() -> TraceMissOptions {
+        Self::default().trace_miss_opts
+    }
+    fn def_trace_miss_correct_max_steps() -> u32 {
+        Self::default().trace_miss_correct_max_steps
+    }
+    fn def_f2_const_air_density() -> bool {
+        Self::default().f2_const_air_density
+    }
+}
+
+impl SerdeAPI for SimParams {}
+impl Init for SimParams {}
+
+impl Default for SimParams {
+    fn default() -> Self {
+        Self {
+            ach_speed_max_iter: 3,
+            ach_speed_tol: 1.0e-3 * uc::R,
+            ach_speed_solver_gain: 0.9,
+            trace_miss_tol: Default::default(),
+            trace_miss_opts: Default::default(),
+            trace_miss_correct_max_steps: 6,
+            f2_const_air_density: true,
+            ambient_thermal_soak: false,
+        }
+    }
+}
+
+#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
+#[serde(deny_unknown_fields)]
+#[non_exhaustive]
+// NOTE: consider embedding this in TraceMissOptions::AllowChecked
+pub struct TraceMissTolerance {
+    /// if the vehicle falls this far behind trace in terms of absolute
+    /// difference and [TraceMissOptions::is_allow_checked], fail
+    pub tol_dist: si::Length,
+    /// if the vehicle falls this far behind trace in terms of fractional
+    /// difference and [TraceMissOptions::is_allow_checked], fail
+    pub tol_dist_frac: si::Ratio,
+    /// if the vehicle falls this far behind instantaneous speed and
+    /// [TraceMissOptions::is_allow_checked], fail
+    pub tol_speed: si::Velocity,
+    /// if the vehicle falls this far behind instantaneous speed in terms of
+    /// fractional difference and [TraceMissOptions::is_allow_checked], fail
+    pub tol_speed_frac: si::Ratio,
+}
+
+impl TraceMissTolerance {
+    pub fn check_trace_miss(
+        &self,
+        cyc_speed: si::Velocity,
+        ach_speed: si::Velocity,
+        cyc_dist: si::Length,
+        ach_dist: si::Length,
+    ) -> anyhow::Result<()> {
+        ensure!(
+            (cyc_speed - ach_speed).abs() < self.tol_speed,
+            "{}\n{}\n{}",
+            format_dbg!(cyc_speed),
+            format_dbg!(ach_speed),
+            format_dbg!(self.tol_speed)
+        );
+        // if condition to prevent divide-by-zero errors
+        if cyc_speed > self.tol_speed {
+            ensure!(
+                (cyc_speed - ach_speed).abs() / cyc_speed < self.tol_speed_frac,
+                "{}\n{}\n{}",
+                format_dbg!(cyc_speed),
+                format_dbg!(ach_speed),
+                format_dbg!(self.tol_speed_frac)
+            )
+        }
+        ensure!(
+            (cyc_dist - ach_dist).abs() < self.tol_dist,
+            "{}\n{}\n{}",
+            format_dbg!(cyc_dist),
+            format_dbg!(ach_dist),
+            format_dbg!(self.tol_dist)
+        );
+        // if condition to prevent checking early in cycle
+        if cyc_dist > self.tol_dist * 5.0 {
+            ensure!(
+                (cyc_dist - ach_dist).abs() / cyc_dist < self.tol_dist_frac,
+                "{}\n{}\n{}",
+                format_dbg!(cyc_dist),
+                format_dbg!(ach_dist),
+                format_dbg!(self.tol_dist_frac)
+            )
+        }
+
+        Ok(())
+    }
+}
+impl SerdeAPI for TraceMissTolerance {}
+impl Init for TraceMissTolerance {}
+impl Default for TraceMissTolerance {
+    fn default() -> Self {
+        Self {
+            tol_dist: 100. * uc::M,
+            tol_dist_frac: 0.05 * uc::R,
+            tol_speed: 10. * uc::MPS,
+            tol_speed_frac: 0.5 * uc::R,
+        }
+    }
+}
+
+#[derive(
+    Clone, Default, Debug, Deserialize, Serialize, PartialEq, IsVariant, derive_more::From, TryInto,
+)]
+pub enum TraceMissOptions {
+    /// Allow trace miss without any fanfare
+    Allow,
+    /// Allow trace miss within error tolerance
+    AllowChecked,
+    #[default]
+    /// Error out when trace miss happens
+    Error,
+    /// Correct trace miss with driver model that catches up
+    Correct,
+}
+
+impl SerdeAPI for TraceMissOptions {}
+impl Init for TraceMissOptions {}
